implement ospl builtin topic
