+++ src/dds_access/idlc.py
@@ -1,0 +1,112 @@
+"""
+ * Copyright(c) 2024 Sven Trittler
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0, or the Eclipse Distribution License
+ * v. 1.0 which is available at
+ * http://www.eclipse.org/org/documents/edl-v10.php.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
+"""
+
+import logging
+import os
+import sys
+from PySide6.QtCore import QDir
+from PySide6.QtCore import QThread, Signal, QFile, QProcess
+import glob
+
+
+class IdlcWorkerThread(QThread):
+
+    doneSignale = Signal()
+    
+    def __init__(self, urls, destination_folder_py, destination_folder_idl, parent=None):
+        super().__init__(parent)
+        self.urls = urls
+        self.destination_folder_idl = destination_folder_idl
+        self.destination_folder_py = destination_folder_py
+
+    def run(self):
+        for url in self.urls:
+            logging.debug("Copy " + str(url) + " ...")
+            if url.isLocalFile():
+                # Copy idl source file
+                source_file = url.toLocalFile()
+                logging.debug("IDL-Folder: " + self.destination_folder_idl)
+                if not QDir(self.destination_folder_idl).exists():
+                    QDir().mkpath(self.destination_folder_idl)
+
+                destination_file = os.path.join(self.destination_folder_idl, os.path.basename(source_file))
+
+                if (QFile.exists(destination_file)):
+                    QFile.remove(destination_file)
+
+                if QFile.copy(source_file, destination_file):
+                    logging.debug("File copied successfully. " + os.path.basename(source_file))
+                else:
+                    logging.error("Failed to copy file.")
+                    break
+
+        parent_dir = self.destination_folder_idl
+        idls = [name for name in os.listdir(parent_dir) if os.path.isfile(os.path.join(parent_dir, name))]
+
+        for idl in idls:
+            logging.debug("Process " + idl + " ...")
+
+            destination_file = os.path.join(self.destination_folder_idl, idl)
+
+            # Compile idl to py file
+            if not QDir(self.destination_folder_py).exists():
+                QDir().mkpath(self.destination_folder_py)
+
+            arguments = ["-l"]
+            application_path = "./"
+
+            if getattr(sys, 'frozen', False):
+                # Bundled as App - use idlc and _idlpy from app binaries
+                application_path = sys._MEIPASS
+                search_pattern = os.path.join(application_path, "_idlpy.*")
+                matching_files = glob.glob(search_pattern)
+                matching_files.sort()
+                if matching_files:
+                    arguments.append(os.path.normpath(matching_files[0]))
+                    logging.debug("Found _idlpy: " + matching_files[0])
+                else:
+                    logging.critical("No _idlpy lib found")
+            else:
+                arguments.append("py")
+                # Started as python program
+                #   - use idlc from cyclonedds_home
+                #   - use _idlpy from pip package
+                if "CYCLONEDDS_HOME" in os.environ:
+                    application_path = os.environ["CYCLONEDDS_HOME"] + "/bin"
+
+            arguments.append("-o")
+            arguments.append(os.path.normpath(self.destination_folder_py))
+            arguments.append("-I")
+            arguments.append(os.path.normpath(self.destination_folder_idl))
+            arguments.append("-f")
+            arguments.append("case-sensitive")
+            arguments.append(os.path.normpath(destination_file))
+
+            command = os.path.normpath(f"{application_path}/idlc")
+
+            logging.info("Execute: " + command + " " + " ".join(arguments))
+
+            process = QProcess()
+            process.setProcessChannelMode(QProcess.ProcessChannelMode.MergedChannels)
+            process.setWorkingDirectory(self.destination_folder_py)
+            process.start(command, arguments)
+
+            if process.waitForFinished():
+                if process.exitStatus() == QProcess.NormalExit:
+                    logging.debug(str(process.readAll()))
+                    logging.debug("Process finished successfully.") 
+                else:
+                    logging.debug("Process failed with error code: " + str(process.exitCode()))
+            else:
+                logging.debug("Failed to start process:" + str(process.errorString()))
+
+        self.doneSignale.emit()
